% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centrePeaks.R
\name{centrePeaks}
\alias{centrePeaks}
\alias{centrePeaks,GRanges,BamFileList-method}
\title{Re-estimate peak centres from coverage}
\usage{
centrePeaks(x, y, ...)

\S4method{centrePeaks}{GRanges,BamFileList}(
  x,
  y,
  f = c("mean", "median", "weighted.mean"),
  BPPARAM = bpparam(),
  ...
)
}
\arguments{
\item{x}{A set of GRanges representing peaks}

\item{y}{A BamFileList}

\item{...}{Not used}

\item{f}{The function to use when estimating a combined peak centre}

\item{BPPARAM}{An object of class BPPARAM}
}
\value{
A GRanges object with all widths set to one
}
\description{
Use coverage to estimate peak centres
}
\details{
Use coverage to estimate the centre of a set of peaks or GenomicRanges.
The point of maximum coverage for each sample will be found and these
positions will be averaged to return a position representing an estimated
peak centre.
If using weighted.mean, positions are weighted by the maximum coverage
within each sample.
}
